requirements.txt
Django==1.5.5
MySQL-python==1.2.5
argparse==1.2.1
django-braces==1.3.1
django-cors-headers==0.12
django-guardian==1.1.1
django-oauth-toolkit==0.5.0
django-registration==1.0
djangorestframework==2.3.7
gunicorn==18.0
oauthlib==0.6.0
python-openid==2.2.5
python-social-auth==0.1.13
requests==1.2.3
requests-oauthlib==0.4.0
shortuuid==0.4
six==1.5.2
wsgiref==0.1.2
